pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["Daniel Ivanov <daniel.ivanov@me.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.scripts]
start = "api.app:run"


[tool.poetry.dependencies]
python = "^3.12"
pydantic = "^2.7.4"
imap-tools = "^1.6.0"
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
apscheduler = "^3.10.4"
supabase = "^2.5.1"
pydantic-settings = "^2.3.3"
pysentimiento = "^0.7.3"
pyjwt = "^2.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
